+++ lib/src/pie_menu_core.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
+import 'package:pie_menu/src/bouncing_widget.dart';
 import 'package:pie_menu/src/pie_action.dart';
 import 'package:pie_menu/src/pie_button.dart';
 import 'package:pie_menu/src/pie_canvas.dart';
@@ -79,8 +80,8 @@ class _PieMenuCoreState extends State<PieMenuCore>
 
   /// Bounce animation for the child widget.
   late final _bounceAnimation = Tween(
-    begin: 1.0,
-    end: _theme.childBounceFactor,
+    begin: 0.0,
+    end: 1.0,
   ).animate(
     CurvedAnimation(
       parent: _bounceController,
@@ -91,6 +92,9 @@ class _PieMenuCoreState extends State<PieMenuCore>
 
   /// Offset of the press event.
   var _pressedOffset = Offset.zero;
+
+  /// Offset of the press event relative to the child widget.
+  var _locallyPressedOffset = Offset.zero;
 
   /// Button used for the press event.
   var _pressedButton = 0;
@@ -104,6 +108,9 @@ class _PieMenuCoreState extends State<PieMenuCore>
   /// Used to measure the time between bounce and debounce.
   final _bounceStopwatch = Stopwatch();
 
+  /// Whether the press was canceled by a pointer move event or menu toggle.
+  var _pressCanceled = false;
+
   /// Controls the shared state.
   PieNotifier get _notifier => PieNotifier.of(context);
 
@@ -115,9 +122,20 @@ class _PieMenuCoreState extends State<PieMenuCore>
   /// If the [PieMenu] does not have a theme, [PieCanvas] theme is used.
   PieTheme get _theme => widget.theme ?? _notifier.canvasTheme;
 
+  /// Whether to use [Listener] instead of [GestureDetector]
+  /// for bounce animation.
+  ///
+  /// [GestureDetector] provides a smoother bounce animation, but its callbacks
+  /// are not triggered when the delay duration is too short because of the
+  /// [LongPressGestureRecognizer] used internally.
+  bool get _useListenerForBounce {
+    return _theme.delayDuration < const Duration(milliseconds: 100);
+  }
+
   @override
   void dispose() {
     _overlayFadeController.dispose();
+    _bounceController.dispose();
     super.dispose();
   }
 
@@ -127,6 +145,7 @@ class _PieMenuCoreState extends State<PieMenuCore>
       if (!_previouslyActive && _state.active) {
         _overlayFadeController.forward(from: 0);
         _debounce();
+        _pressCanceled = true;
       } else if (_previouslyActive && !_state.active) {
         _overlayFadeController.reverse();
       }
@@ -140,34 +159,48 @@ class _PieMenuCoreState extends State<PieMenuCore>
 
     return Stack(
       children: [
-        Positioned.fill(
-          child: AnimatedBuilder(
-            animation: _overlayFadeAnimation,
-            builder: (context, child) {
-              return Opacity(
-                opacity: _overlayFadeAnimation.value,
-                child: child,
-              );
-            },
-            child: ColoredBox(color: _theme.effectiveOverlayColor),
+        if (_theme.overlayStyle == PieOverlayStyle.around)
+          Positioned.fill(
+            child: AnimatedBuilder(
+              animation: _overlayFadeAnimation,
+              builder: (context, child) {
+                return Opacity(
+                  opacity: _overlayFadeAnimation.value,
+                  child: child,
+                );
+              },
+              child: ColoredBox(color: _theme.effectiveOverlayColor),
+            ),
           ),
-        ),
         MouseRegion(
           cursor: SystemMouseCursors.click,
           child: Listener(
             onPointerDown: _pointerDown,
             onPointerMove: _pointerMove,
             onPointerUp: _pointerUp,
-            child: AnimatedBuilder(
-              animation: _bounceAnimation,
-              builder: (context, child) {
-                return Transform.scale(
-                  scale: _bounceAnimation.value,
-                  alignment: Alignment.center,
-                  child: child,
-                );
-              },
-              child: widget.child,
+            child: GestureDetector(
+              onTapDown: _tapDown,
+              onTapCancel: _tapCancel,
+              onTapUp: _tapUp,
+              dragStartBehavior: DragStartBehavior.down,
+              child: AnimatedOpacity(
+                opacity: _theme.overlayStyle == PieOverlayStyle.around &&
+                        _state.menuKey == _uniqueKey &&
+                        _state.active &&
+                        _state.hoveredAction != null
+                    ? _theme.childOpacityOnButtonHover
+                    : 1,
+                duration: _theme.hoverDuration,
+                curve: Curves.ease,
+                child: _theme.childBounceEnabled
+                    ? BouncingWidget(
+                        theme: _theme,
+                        animation: _bounceAnimation,
+                        locallyPressedOffset: _locallyPressedOffset,
+                        child: widget.child,
+                      )
+                    : widget.child,
+              ),
             ),
           ),
         ),
@@ -175,13 +208,30 @@ class _PieMenuCoreState extends State<PieMenuCore>
     );
   }
 
-  void _pointerDown(PointerDownEvent event) async {
-    _pressedOffset = event.position;
-    _pressedButton = event.buttons;
+  void _tapDown(TapDownDetails details) {
+    if (!_useListenerForBounce) _bounce();
+  }
+
+  void _tapCancel() {
+    if (!_useListenerForBounce) _debounce();
+  }
+
+  void _tapUp(TapUpDetails details) {
+    if (!_useListenerForBounce) _debounce();
+  }
+
+  void _pointerDown(PointerDownEvent event) {
+    setState(() {
+      _pressedOffset = event.position;
+      _locallyPressedOffset = event.localPosition;
+      _pressedButton = event.buttons;
+    });
 
     if (_state.active) return;
 
-    _bounce();
+    _pressCanceled = false;
+
+    if (_useListenerForBounce) _bounce();
 
     final isMouseEvent = event.kind == PointerDeviceKind.mouse;
     final leftClicked = isMouseEvent && _pressedButton == kPrimaryMouseButton;
@@ -197,7 +247,10 @@ class _PieMenuCoreState extends State<PieMenuCore>
     _notifier.canvas.attachMenu(
       rightClicked: rightClicked,
       offset: _pressedOffset,
+      localOffset: _locallyPressedOffset,
       renderBox: context.findRenderObject() as RenderBox,
+      child: widget.child,
+      bounceAnimation: _bounceAnimation,
       menuKey: _uniqueKey,
       actions: widget.actions,
       theme: _theme,
@@ -215,16 +268,15 @@ class _PieMenuCoreState extends State<PieMenuCore>
     if (_state.active) return;
 
     if ((_pressedOffset - event.position).distance > 8) {
-      _debounce();
+      _pressCanceled = true;
+      if (_useListenerForBounce) _debounce();
     }
   }
 
   void _pointerUp(PointerUpEvent event) {
-    _debounce();
+    if (_useListenerForBounce) _debounce();
 
-    if ((_pressedOffset - event.position).distance > 8) {
-      return;
-    }
+    if (_pressCanceled) return;
 
     if (_state.active && _theme.delayDuration != Duration.zero) {
       return;
@@ -240,13 +292,13 @@ class _PieMenuCoreState extends State<PieMenuCore>
   }
 
   void _bounce() {
-    if (!_theme.childBounceEnabled) return;
-
-    _bounceController.forward();
+    if (!_theme.childBounceEnabled || _bounceStopwatch.isRunning) return;
 
     _debounceTimer?.cancel();
     _bounceStopwatch.reset();
     _bounceStopwatch.start();
+
+    _bounceController.forward();
   }
 
   void _debounce() {
@@ -254,9 +306,11 @@ class _PieMenuCoreState extends State<PieMenuCore>
 
     _bounceStopwatch.stop();
 
-    final debounceDelay = _bounceStopwatch.elapsedMilliseconds > 100
+    final minDelayMS = _theme.delayDuration == Duration.zero ? 100 : 75;
+
+    final debounceDelay = _bounceStopwatch.elapsedMilliseconds > minDelayMS
         ? Duration.zero
-        : const Duration(milliseconds: 100);
+        : Duration(milliseconds: minDelayMS);
 
     _debounceTimer = Timer(debounceDelay, () {
       if (mounted) _bounceController.reverse();